Rename project files and directories to panos_device_certificate_report
This commit renames all project files and directories from device_certificate_report to panos_device_certificate_report for better project structure clarity. It includes updates to import statements, the README file examples, and the pyproject.toml configuration.
