+++ panos_device_certificate_report/components/__init__.py
@@ -1,8 +1,7 @@
-# components/__init__.py
+# panos_device_certificate_report/components/__init__.py
 
 from .data_collection import (
     process_csv_file,
     collect_data_from_panorama,
     collect_data_from_firewall,
 )
-from .report_generation import generate_report
