+++ panos_device_certificate_report/models/device.py
