+++ panos_device_certificate_report/utilities/filters.py
@@ -1,13 +1,10 @@
-# device_certificate_report/components/filters.py
+# panos_device_certificate_report/components/filters.py
 
 from typing import List, Tuple
-from device_certificate_report.components.config import (
-    AffectedModels,
-    UnaffectedModels,
-    MinimumPatchedVersions,
-)
-from device_certificate_report.models.device import DeviceInfo
-from device_certificate_report.components.version import parse_version
+from panos_device_certificate_report.config.hardware_families import AffectedModels, UnaffectedModels
+from panos_device_certificate_report.config.panos_versions import MinimumPatchedVersions
+from panos_device_certificate_report.models.device import DeviceInfo
+from panos_device_certificate_report.components.version import parse_version
 
 
 def is_affected_model(model: str) -> bool:
